Setup Database
Added scripts and methods to setup and drop the database schema. If you run server, open the login page, and click the Login button it will run the method to create the tables. Added sql files for setting up the tables, dropping the tables, and one for a clearer understanding of the table schema.